UC-12 refactor: /entities/PolicySwitcher via FSD methodology
