+++ src/entities/PolicySwitcher/index.ts
@@ -1,0 +1,1 @@
+export { PolicySwitcher } from './ui/PolicySwitcher'
