+++ src/entities/PolicySwitcher/model/types.ts
@@ -1,0 +1,3 @@
+export type PolicySwitcherProps = {
+  variant: 'policy' | 'terms'
+}
